Refactor window.rs
